Move dispatcher to use a processor
